services/led-daemon/requirements.txt
Adafruit-Blinka==8.58.1
adafruit-circuitpython-busdevice==5.2.12
adafruit-circuitpython-connectionmanager==3.1.4
adafruit-circuitpython-motor==3.4.16
adafruit-circuitpython-pca9685==3.4.18
adafruit-circuitpython-register==1.10.3
adafruit-circuitpython-requests==4.1.12
adafruit-circuitpython-typing==1.12.1
Adafruit-GPIO==1.0.3
adafruit-io==2.8.0
Adafruit-PCA9685==1.0.1
Adafruit-PlatformDetect==3.79.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
board==1.0
cbor2==5.6.5
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
idna==3.10
lgpio==0.2.2.0
luma==2.3.20
luma.core==2.4.2
luma.oled==3.14.0
paho-mqtt==2.1.0
pillow==11.2.1
pycparser==2.21
pyftdi==0.56.0
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.0
pyusb==1.3.1
requests==2.32.3
RPi.GPIO>=0.7.0
six==1.17.0
smbus2==0.5.0
spidev==3.7
sysv-ipc==1.1.0
typing_extensions==4.14.0
urllib3==2.4.0
mcp==1.9.3
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
